refactor(server): split team.ts — extract 3 mutation handlers (S27.8.25)
Fourth slice on `apps/server/src/routes/team.ts`. Extract the 3
team mutation handlers into a dedicated module
`routes/team-mutation-handlers.ts` (304 lines, under DoD 400).

Endpoints covered:
- `PUT /:id/info` — handlePutTeamInfo (inducements: rerolls,
  cheerleaders, assistants, apothecary, dedicatedFans, with
  match-in-progress lock)
- `POST /:id/recalculate` — handleRecalculateTeam (force TV/CV
  recompute via updateTeamValues)
- `PUT /:id` — handleUpdateTeam (rename team + update
  player names/numbers transactional, with validation)

All thematically cohesive (team-level mutation + roster cosmetic
update). Re-export from `team.ts` to preserve the public API
consumed by `team.test.ts`.

Leaf dependencies only (`prisma`, `Prisma.PrismaPromise`,
`sendError/sendSuccess`, `updateTeamValues`, `serverLog`). No cycle.

Cleanup of orphaned imports in `team.ts` (1 removed): `Prisma`
from `@prisma/client`.

`team.ts` shrinks 1485 -> 1257 lines (-228 net lines, cumulative
-1157 vs origin 2414 since S27.8.22 = -48%).

5 new smoke tests `team-mutation-handlers.test.ts`:
- exports x2 (named exports + re-export from team.ts)
- 404 gates on team not found x3

- 1819/1819 server tests pass (was 1814, +5 new)
- typecheck clean
